precision update

+++ myEnergiConnect/Model/External/Shared/IMyEnergiProductWithCt.cs
@@ -5,7 +5,7 @@ public interface IMyEnergiProductWithCt : IMyEnergiProduct
     string Ct1Name { get; init; }
     string Ct2Name { get; init; }
     string Ct3Name { get; init; }
-    double PhysicalCt1Value { get; init; }
-    double PhysicalCt2Value { get; init; }
-    double PhysicalCt3Value { get; init; }
+    decimal PhysicalCt1Value { get; init; }
+    decimal PhysicalCt2Value { get; init; }
+    decimal PhysicalCt3Value { get; init; }
 }
